consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["Amineh Dadsetan <amineh.dadsetan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyspark = "^3.5.4"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
